requirements.txt
fastapi[all]
uvicorn
sqlalchemy
bcrypt
passlib
psycopg2
python-magic-bin
cloudinary
python-jose[cryptography]
pydantic-settings
alembic